refactor: add input.rs and output.rs
