+++ src/client.rs
@@ -1,7 +1,4 @@
-use crate::{
-    error::AppError,
-    routes::{Input, Output},
-};
+use crate::{error::AppError, input::routes::Input, output::routes::Output};
 
 pub trait GoogleMapClient {
     fn routes(&self, input: Input) -> Result<Output, AppError>;
